+++ HiKiApp/Resource/Base.lproj/LaunchScreen.storyboard
@@ -17,13 +17,9 @@
                         <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
-                            <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="splash.png" translatesAutoresizingMaskIntoConstraints="NO" id="N3v-pR-xoq">
-                                <rect key="frame" x="-5" y="-5" width="403" height="862"/>
-                                <color key="tintColor" name="pointColor"/>
-                            </imageView>
                             <view alpha="0.29999999999999999" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="Uce-6i-5nG">
                                 <rect key="frame" x="-5" y="-5" width="403" height="862"/>
-                                <color key="backgroundColor" white="0.0" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                                <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                             </view>
                             <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="clearLogo.png" translatesAutoresizingMaskIntoConstraints="NO" id="shq-H2-EFF">
                                 <rect key="frame" x="24" y="231" width="345" height="150"/>
@@ -37,19 +33,13 @@
                         <color key="tintColor" name="customWhite"/>
                         <constraints>
                             <constraint firstItem="Uce-6i-5nG" firstAttribute="centerX" secondItem="Ze5-6b-2t3" secondAttribute="centerX" id="0Mq-p0-BDT"/>
-                            <constraint firstItem="N3v-pR-xoq" firstAttribute="centerY" secondItem="Ze5-6b-2t3" secondAttribute="centerY" id="0uh-Nu-Hgd"/>
-                            <constraint firstItem="6Tk-OE-BBY" firstAttribute="trailing" secondItem="N3v-pR-xoq" secondAttribute="trailing" constant="-5" id="1PF-BT-siQ"/>
-                            <constraint firstItem="N3v-pR-xoq" firstAttribute="top" secondItem="Ze5-6b-2t3" secondAttribute="top" constant="-5" id="7cH-88-yIl"/>
-                            <constraint firstItem="N3v-pR-xoq" firstAttribute="centerX" secondItem="Ze5-6b-2t3" secondAttribute="centerX" id="LMh-bF-UMQ"/>
                             <constraint firstItem="shq-H2-EFF" firstAttribute="leading" secondItem="6Tk-OE-BBY" secondAttribute="leading" constant="24" id="QXj-5h-ffp"/>
                             <constraint firstItem="6Tk-OE-BBY" firstAttribute="trailing" secondItem="shq-H2-EFF" secondAttribute="trailing" constant="24" id="SiY-hF-u2K"/>
                             <constraint firstItem="Uce-6i-5nG" firstAttribute="leading" secondItem="6Tk-OE-BBY" secondAttribute="leading" constant="-5" id="Ton-5d-C5K"/>
                             <constraint firstItem="shq-H2-EFF" firstAttribute="centerY" secondItem="Ze5-6b-2t3" secondAttribute="centerY" constant="-120" id="UMC-vv-PYt"/>
                             <constraint firstItem="Uce-6i-5nG" firstAttribute="centerY" secondItem="Ze5-6b-2t3" secondAttribute="centerY" id="ebA-6L-Wp3"/>
-                            <constraint firstAttribute="bottom" secondItem="N3v-pR-xoq" secondAttribute="bottom" constant="-5" id="eo3-JE-eQo"/>
                             <constraint firstAttribute="bottom" secondItem="Uce-6i-5nG" secondAttribute="bottom" constant="-5" id="k3s-bj-2tv"/>
                             <constraint firstItem="6Tk-OE-BBY" firstAttribute="trailing" secondItem="Uce-6i-5nG" secondAttribute="trailing" constant="-5" id="qIc-Ri-fmF"/>
-                            <constraint firstItem="N3v-pR-xoq" firstAttribute="leading" secondItem="6Tk-OE-BBY" secondAttribute="leading" constant="-5" id="r9A-cn-dPM"/>
                             <constraint firstItem="Uce-6i-5nG" firstAttribute="top" secondItem="Ze5-6b-2t3" secondAttribute="top" constant="-5" id="t3Z-03-Vuz"/>
                         </constraints>
                     </view>
@@ -61,12 +51,8 @@
     </scenes>
     <resources>
         <image name="clearLogo.png" width="1024" height="600"/>
-        <image name="splash.png" width="4096" height="2734"/>
         <namedColor name="customWhite">
             <color red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-        </namedColor>
-        <namedColor name="pointColor">
-            <color red="0.96078431372549022" green="0.56470588235294117" blue="0.28235294117647058" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
         </namedColor>
     </resources>
 </document>
